Move key handling to on_key, use Vanilla keys where possible
Parsing escape codes ourselves to figure out which keys are pressed is too much
work (and the current code was likely incomplete, as the reported sequence can
probably change depending on terminal settings?)

Also rely on the Vanilla key mappings where possible.
Remove the sprint/strafe toggle as Shift/Alt is actually not _terrible_ in the
terminal.