Initial implementation of detached_ui DOOM config var
To be used when using the typical cell-based gfx backend.

As most UI text is illegible in that mode, we'll use this config var to hide
most UI elements and instead draw our own.

Demo this with a new msg type for getting HUD text without rendering it. It's
likely we'll be moving the cell gfx code to the executable anyway, for
performance reasons.