Initial config module implementation
Also set the cwd of the DOOM process to the executable dir. Important for save
games and such (though there's also a config var for that).

TODO: Need to figure out what to put in it, but also the merging behaviour; how
should we specify fields that we want to set to nil? nil won't work, as that
is treated as a non-existent field. vim.NIL? Avoid nils? Idk