Fix coroutine yielding for PUC 5.1
Can't yield across a pcall in PUC, as it's a C call.

This does mean there'll be no way to get a backtrace from errors within the
coroutines anymore (I was planning on using an xpcall handler for that), but oh
well.

Also clarify that LuaJIT support is recommended, though not required.